refactor: move name validation to separate file
exporting validation files and regex consts
