+++ packages/amplify-cli/src/extensions/amplify-helpers/project-name-validation.js
@@ -1,0 +1,30 @@
+const { makeId } = require('./make-id');
+
+const validAlphanumericRegex = /^[a-zA-Z0-9]+$/g;
+const invalidAlphanumericRegex = /[^a-zA-Z0-9]/g;
+
+function isProjectNameValid(projectName) {
+  return projectName && projectName.length >= 3 && projectName.length <= 20 && validAlphanumericRegex.test(projectName);
+}
+
+function normalizeProjectName(projectName) {
+  if (!projectName) {
+    projectName = `amplify${makeId(5)}`;
+  }
+  if (!isProjectNameValid(projectName)) {
+    projectName = projectName.replace(invalidAlphanumericRegex, '');
+    if (projectName.length < 3) {
+      projectName += makeId(5);
+    } else if (projectName.length > 20) {
+      projectName = projectName.substring(0, 20);
+    }
+  }
+  return projectName;
+}
+
+module.exports = {
+  validAlphanumericRegex,
+  invalidAlphanumericRegex,
+  isProjectNameValid,
+  normalizeProjectName,
+};
